pipecat-agent/requirements.txt
pipecat-ai[deepgram,openai,silero]
pipecat-murf-tts
fastapi
uvicorn[standard]
websockets
aiohttp
python-dotenv
twilio
